% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angles.R
\name{angles}
\alias{angles}
\title{angle between two spaces}
\usage{
angles(B1, B2)
}
\arguments{
\item{B1}{The name of first space}

\item{B2}{The name of second space}
}
\description{
angle between two spaces
}
