% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta.order.R
\name{beta.order}
\alias{beta.order}
\title{choose the order of beta}
\usage{
beta.order(beta)
}
\arguments{
\item{beta}{the result beta of sparse sufficient dimension reduction}
}
\description{
choose the order of beta
}
