% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sSDR.R
\name{center}
\alias{center}
\title{Center a vector}
\usage{
center(v)
}
\arguments{
\item{v}{A vector.}
}
\value{
A vector with mean zero.
}
\description{
Center a vector
}
\details{
This function centers any vector and returns a vector with mean zero.
}
\examples{
data <- gen.data(n=100)
y.centered <- center(data$y)
}

