% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.sdr.R
\name{comp.sdr}
\alias{comp.sdr}
\title{Compute with sufficient dimension reduction method}
\usage{
comp.sdr(X, y, method, d, nslices)
}
\arguments{
\item{X}{The name of predictor matrix}

\item{y}{The name of response matrix}

\item{method}{sufficient dimension reduction method}

\item{d}{The number of dimensions to reduce}

\item{nslices}{The number of slices}
}
\value{
The output from sufficient dimension reduction
}
\description{
Compute with sufficient dimension reduction method
}
