% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sSDR.R
\name{cov.x}
\alias{cov.x}
\title{Covariance matrix}
\usage{
cov.x(X)
}
\arguments{
\item{X}{a n x p matrix of n observations and p predictors.}
}
\value{
A p x p covariance matrix.
}
\description{
Covariance matrix
}
\details{
This function returns A p x p covariance matrix for any n x p matrix.
}
\examples{
data <- gen.data(n=100)
x.cov <- cov.x(data$X)
}

