% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{dist}
\alias{dist}
\title{distance between subspaces}
\usage{
dist(v1, v2)
}
\arguments{
\item{v1}{The name of first subspace}

\item{v2}{The name of second subspace}
}
\description{
distance between subspaces
}
