% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.sqrt.inv.R
\name{mat.sqrt.inv}
\alias{mat.sqrt.inv}
\title{square-root-inverse of a matrix}
\usage{
mat.sqrt.inv(A)
}
\arguments{
\item{A}{The name of matrix to find square-root-inverse of matrix}
}
\description{
square-root-inverse of a matrix
}
