% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sSDR.R
\name{matpower}
\alias{matpower}
\title{Power of a matrix}
\usage{
matpower(X, alpha)
}
\arguments{
\item{X}{A p x p square matrix.}

\item{alpha}{A scaler determining the order of the power.}
}
\value{
A p x p square matrix.
}
\description{
Power of a matrix
}
\details{
This function calculates the power of a square matrix.
}
\examples{
data <- gen.data(n=100)
cov.squared <- matpower(cov.x(data$X), 2)
}

