% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sSDR.R
\name{norm1}
\alias{norm1}
\title{Normalize a vector}
\usage{
norm1(v)
}
\arguments{
\item{v}{A vector.}
}
\value{
A vector with norm 1.
}
\description{
Normalize a vector
}
\details{
This function normalizes any non-zero vector and returns a vector with
the norm equal to 1.
}
\examples{
data <- gen.data(n=100)
y.norm1 <- norm1(data$y)
}

