% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sSDR.R
\name{orthnormal}
\alias{orthnormal}
\title{Gram-Schmidt orthonormalization}
\usage{
orthnormal(X)
}
\arguments{
\item{X}{a n x p matrix of n observations and p predictors.}
}
\value{
A n x p matrix of n observations and p predictors.
}
\description{
Gram-Schmidt orthonormalization
}
\details{
This function orthonormalizes any n x p matrix.
}
\examples{
data <- gen.data(n=100)
x.orth <- orthnormal(data$X)
}

