% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{print.ssdr_fit}
\alias{print.ssdr_fit}
\title{Print an SSDr Fit}
\usage{
\method{print}{ssdr_fit}(x, ...)
}
\arguments{
\item{x}{An \code{ssdr_fit} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an SSDr Fit
}
