% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_beta.R
\name{solve_beta}
\alias{solve_beta}
\title{solve beta using lasso estimation}
\usage{
solve_beta(x, y, G2, lambda1, lambda2)
}
\arguments{
\item{x}{The name of predictor matrix}

\item{y}{The name of response matrix}

\item{G2}{The square root of covariance x}

\item{lambda1}{shrinkage parameters}

\item{lambda2}{shrinkage parameters}
}
\description{
solve beta using lasso estimation
}
