% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssdr.wrap.R
\name{ssdr.wrap}
\alias{ssdr.wrap}
\title{Function to get optimal lambda}
\usage{
ssdr.wrap(
  X,
  y,
  method = c("pc", "sir", "save", "phdres", "dr"),
  d,
  nslices,
  s1.range,
  s2.range,
  max.iter,
  eps.conv
)
}
\arguments{
\item{X}{The name of predictor matrix}

\item{y}{The name of response matrix}

\item{method}{sufficient dimension reduction method}

\item{d}{The number of dimensions to reduce}

\item{nslices}{The number of slices}

\item{s1.range}{shrinkage parameters}

\item{s2.range}{shrinkage parameters}

\item{max.iter}{The number of max iterations}

\item{eps.conv}{Epsilon convergence}
}
\value{
Provide optimal lambda
}
\description{
Function to get optimal lambda
}
