% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sSDR.R
\name{standmat}
\alias{standmat}
\title{Matrix standardization}
\usage{
standmat(x)
}
\arguments{
\item{x}{A n x p matrix of n observations and p predictors.}
}
\value{
A n x p matrix of n observations and p predictors.
}
\description{
Matrix standardization
}
\details{
This function standardizes a matrix treating each row as a random vector
in an iid sample. It returns a n x p matrix with column-mean zero
and identity-covariance matrix.
}
\examples{
data <- gen.data(n=100)
x.std <- standmat(data$X)
}

